pyproject.toml
[tool.poetry]
name = "sensor-data-migrator"
version = "0.1.0"
description = ""
authors = ["Prakash Prajapati <prakash.prajapati@tekkon.com.np>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^1.4.45"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.0"
boto3 = "^1.26.37"
boto3-stubs = {extras = ["dynamodb"], version = "^1.34.62"}


[tool.poetry.dev-dependencies]
black = "^22.12.0"
pytest = "^7.2.0"
ipdb = "^0.13.11"
pre-commit = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
